requirement.txt
APScheduler==3.0.5
attrs==15.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
bootstrap-admin==0.3.6
certifi==2016.2.28
cffi==1.5.2
chardet==2.3.0
click==6.3
cryptography==1.2.3
cssselect==0.9.1
cycler==0.10.0
Cython==0.23.4
Django==1.9.4
django-contrib-comments==1.7.0
django-crispy-forms==1.6.0
django-grappelli==2.8.1
django-markdown==0.8.4
django-pagedown==0.1.1
django-threadedcomments==1.0.1
django-xadmin==0.5.0
DjangoUeditor==1.8.143
dnet==1.12
dpkt==1.7
enum34==1.1.2
Flask==0.10.1
Flask-Login==0.3.2
flup==1.0.2
futures==3.0.5
html2text==2016.4.2
html5lib==0.9999999
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
lxml==3.4.4
Markdown==2.6.6
MarkupSafe==0.23
matplotlib==1.5.1
MySQL-python==1.2.5
numpy==1.9.2
pcap==1.1
Pillow==3.1.1
py2exe==0.6.9
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycurl==7.43.0
Pygments==2.1.3
pyHook==1.5.1
PyMySQL==0.7.2
pyOpenSSL==0.15.1
pyparsing==2.1.0
pyquery==1.2.11
pyspider==0.3.6
python-dateutil==2.5.0
pytz==2015.7
pywin32==219
queuelib==1.4.2
requests==2.9.1
scapy==2.2.0
scipy==0.15.1
Scrapy==1.0.5
selenium==2.53.1
service-identity==16.0.0
singledispatch==3.4.0.3
six==1.10.0
termcolor==1.1.0
tornado==4.3
Twisted==16.0.0
tzlocal==1.2.2
u-msgpack-python==2.1
vboxapi==1.0
virtualenv==15.0.0
w3lib==1.13.0
Werkzeug==0.11.4
WsgiDAV==1.2.0
zope.interface==4.1.3
